vtctl.generateShardRanges -> key.GenerateShardRanges
